Inline assembly

+++ src/vmStructs.cpp
@@ -643,7 +643,7 @@ int VMThread::osThreadId() {
     const char* osthread = *(const char**) at(_thread_osthread_offset);
     if (osthread != NULL) {
         // Java thread may be in the middle of termination, and its osthread structure just released
-        return SafeAccess::load32((u32*)(osthread + _osthread_id_offset), (u32)-1);
+        return SafeAccess::load32((int32_t*)(osthread + _osthread_id_offset), -1);
     }
     return -1;
 }
